HI3516EV200 PLL  initialization
So, I down sure is it real registers, or that was fake description.
I don't know is this memory is only MMU mapped memory with memory fault update access, but I disable all MMU,
and move HYP, USR, SVC, MON and other handler to my boot code. It is not included into github. 
But, for now  I am testing Register access, because I wanted to use it in HiSilicon project, and my one project.

This description will allow to use register safer, and understand each step of initialization or driver work process.